requirements.txt
fastapi
openai
uvicorn
python-multipart
pydantic
pyyaml
pillow
python-dotenv
loguru
colorama
rich
httpx
psycopg2-binary
pgvector
datetime
scikit-learn
numpy
matplotlib
python-multipart
